+++ src/components/Modal/Modal.test.tsx
@@ -3,6 +3,7 @@ import userEvent from '@testing-library/user-event';
 import { runCommonTests } from '../../utils/testUtils';
 import Modal from './Modal';
 import { ModalFromDrawer } from './Modal.stories';
+import { px } from '../../utils';
 
 describe('Modal', () => {
   const onCloseMock = vi.fn();
@@ -86,5 +87,43 @@ describe('Modal', () => {
     const overlay = screen.getByTestId(/overlay/);
     await userEvent.click(overlay);
     expect(onCloseMock).toHaveBeenCalled();
+  });
+
+  it('renders custom close icon', () => {
+    const customCloseIcon = <span>Custom Close</span>;
+    render(
+      <Modal isOpen onClose={onCloseMock} closeIcon={customCloseIcon} closeIconPosition="left">
+        <div>Modal Content</div>
+      </Modal>,
+    );
+
+    const closeButton = screen.getByText('Custom Close');
+    expect(closeButton).toBeInTheDocument();
+  });
+
+  it('renders class to align close icon to the left', () => {
+    const customCloseIcon = <span>Custom Close</span>;
+    render(
+      <Modal isOpen onClose={onCloseMock} closeIcon={customCloseIcon} closeIconPosition="left">
+        <div>Modal Content</div>
+      </Modal>,
+    );
+
+    const closeButton = screen.getByText('Custom Close');
+    expect(closeButton).toBeInTheDocument();
+    expect(closeButton.parentElement).toHaveClass(`${px}-modal__close--left`);
+  });
+
+  it('renders component without left class alignment', () => {
+    const customCloseIcon = <span>Custom Close</span>;
+    render(
+      <Modal isOpen onClose={onCloseMock} closeIcon={customCloseIcon}>
+        <div>Modal Content</div>
+      </Modal>,
+    );
+
+    const closeButton = screen.getByText('Custom Close');
+    expect(closeButton).toBeInTheDocument();
+    expect(closeButton.parentElement).not.toHaveClass(`${px}-modal__close--left`);
   });
 });